refactor(ai-core): split QueryControls into focused modules
`QueryControls.tsx` was 591 lines changing for four unrelated reasons: legacy
`children` routing, composer layout, credential entry, and drag-and-drop. Move
each into its own module under `queryControls/`, leaving the entry file at 156
lines holding only the props, the runtime-mode dispatch, and the two mode
bodies.

- `deprecatedChildren.tsx` — the identity-check `children` routing, whole and
  in one place so the future break is a file deletion plus one call site.
- `InlineApiKeyComposer.tsx` — the credential screen. It had its own state,
  focus effect, keymap, and save path, and shared only footer markup with the
  composer. Its Enter handler and save button now run one `save()` instead of
  two copies of the same validate-and-clear, and `formatProviderLabel` moves to
  module scope instead of being rebuilt each render.
- `ComposerFrame.tsx` — the box, textarea, and footer.
- `ComposerTopRow.tsx`, `ComposerFooterStrip.tsx` — the footer strip's four
  nested divs were duplicated verbatim between the composer and the API-key
  screen. `actions` renders unwrapped, since the two space their actions
  differently.
- `ContextDropTarget.tsx` — the dnd-kit block, isolated for the same reason as
  the sniffing code: it duplicates the `DropTarget` primitive and should
  collapse into it once the recipe styles off `data-drop-active`.
- `useDelayedFocus.ts` — the 500ms focus timeout, previously written twice.

No behavior or markup change: the rendered tree, class names, and every
component interface are unchanged.
